Replace awesome-typescript-loader with ts-loader
awesome-typescript-loader hasn't seen a new release since 2018 and
doesn't work with modern TypeScript versions etc. The GitHub repo was
archived on 2019-12-01.

Using ts-loader works great in Node 16, but in Node 18, a lot of test
cases fail with this error:

    callback(): The callback was already called.

      123 |         callback(null, source);
      124 |     }).catch(err => {
    > 125 |         callback(err, source);
          |         ^
      126 |     });
      127 | }
      128 |

      at context.callback (node_modules/loader-runner/lib/LoaderRunner.js:106:10)
      at src/loader.ts:125:9

But the tests didn't pass in Node 18 with awesome-typescript-loader
either, so … ¯\_(ツ)_/¯

I believe this is the reason by the way:
https://github.com/webpack/webpack/issues/14532